WIP commit before a potential break cause by work assignments.
