+++ src/matching_results/mod.rs
@@ -1,1 +1,1 @@
-pub mod matching_line;
+pub mod result;
